chore: move create category controller to dedicated folder
